Rename event log merging logic
