+++ casper/src/main/scala/coop/rchain/casper/rholang/syntax/RuntimeSyntax.scala
@@ -50,7 +50,7 @@ import coop.rchain.rholang.interpreter.merging.RholangMergingLogic
 import coop.rchain.rholang.interpreter.{storage, EvaluateResult, RhoRuntime}
 import coop.rchain.rspace.hashing.{Blake2b256Hash, StableHashProvider}
 import coop.rchain.rspace.history.History.emptyRootHash
-import coop.rchain.rspace.merger.MergingLogic.NumberChannelsEndVal
+import coop.rchain.rspace.merger.EventLogMergingLogic.NumberChannelsEndVal
 import coop.rchain.shared.{Base16, Log}
 
 trait RuntimeSyntax {
