+++ rspace/src/test/scala/coop/rchain/rspace/merging/EventsMergingLogicSpec.scala
@@ -1,10 +1,10 @@
 package coop.rchain.rspace.merging
 
-import coop.rchain.rspace.merger.MergingLogic._
+import coop.rchain.rspace.merger.EventLogMergingLogic._
 import coop.rchain.shared.Stopwatch
 import org.scalatest.{FlatSpec, Matchers}
 
-class MergingLogicSpec extends FlatSpec with Matchers {
+class EventsMergingLogicSpec extends FlatSpec with Matchers {
   // some random conflict maps and rejection options, computed manually
   "rejection options" should "be computed correctly" in {
     computeRejectionOptions(
